tests/wpt/tests/svg/scripted/script-runs-in-shadow-tree.html

<!doctype html>
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1555949">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="host"></div>
<script>
  let svgScript = document.createElementNS("http://www.w3.org/2000/svg", "script");
  svgScript.innerHTML = "window.scriptRan = true";
  document.getElementById("host").attachShadow({ mode: "open" }).appendChild(svgScript);
  test(function() {
    assert_true(window.scriptRan);
  }, "svg:script runs correctly in a shadow tree");
</script>
